+++ SFML Game/source/Components/Spawner/AsteroidSpawner.cpp
@@ -28,13 +28,13 @@ void AsteroidSpawner::Update(GameObject* gameobject, sf::Int32 deltaTime) {
 
 			// Generate random initial position outside bounds
 			Vec3<float> position;
-			position.SetDirection(RandInt(1, 360));
-			position.SetMagnitude(hypot(_bounds.x, _bounds.y));
+			position.SetDirection((float)RandInt(1, 360));
+			position.SetMagnitude((float)hypot(_bounds.x, _bounds.y));
 
 			// Generate random initial velocity
 			Vec3<float> velocity;
 			velocity.SetMagnitude(RandFloat(MIN_SPEED, MAX_SPEED));
-			velocity.SetDirection(RandInt(1, 360));
+			velocity.SetDirection((float)RandInt(1, 360));
 
 			// Set asteroid to use calculated position and velocity
 			newAsteroid->_sprite.setPosition(_bounds.x / 2.0f + position.GetX(), _bounds.y / 2.0f + position.GetY());
